Refactor level loading and handling
By simplifying the way level loading happens, I realized that the way
map storage in memory was very weird and not clear at all. Took the
chance to refactor everything into clearer names.

In the middle of this, also realized the bug of disapearing monster
corpes is related to this. During the changes, for a moment, the
monsters corpses began appearing.